% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{as_args_list}
\alias{as_args_list}
\title{Get Arguments as List}
\usage{
as_args_list(...)
}
\arguments{
\item{...}{Any number of arguments.}
}
\value{
A list containing the arguments provided.
}
\description{
This function takes a variable number of arguments and
 returns them as a list.
}
\examples{
# Get arguments as a list
args_list <- as_args_list(10, "hello", TRUE)
print(args_list)

}
