% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveobjects.R
\name{savewidget}
\alias{saveobjects}
\alias{savewidget}
\title{RStudio Addin to Save Object from Environment to File}
\usage{
saveobjects()
}
\description{
RStudio Addin to Save Object from Environment to File
}
\examples{
\dontrun{saveobjects()}
}

