% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{Save}
\alias{Save}
\title{Save Class}
\description{
This R6 class provides methods to handle file saving and naming operations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dir}}{The directory path where files will be saved.
Default is the current working directory.}

\item{\code{prefix}}{The prefix used in the file naming. Default is "file".}

\item{\code{count}}{The numeric count used in the file naming. Default is 1.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Save-new}{\code{Save$new()}}
\item \href{#method-Save-make_filename}{\code{Save$make_filename()}}
\item \href{#method-Save-set_dir}{\code{Save$set_dir()}}
\item \href{#method-Save-set_prefix}{\code{Save$set_prefix()}}
\item \href{#method-Save-set_count}{\code{Save$set_count()}}
\item \href{#method-Save-csv}{\code{Save$csv()}}
\item \href{#method-Save-text}{\code{Save$text()}}
\item \href{#method-Save-plot}{\code{Save$plot()}}
\item \href{#method-Save-gg}{\code{Save$gg()}}
\item \href{#method-Save-clone}{\code{Save$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Save-new"></a>}}
\if{latex}{\out{\hypertarget{method-Save-new}{}}}
\subsection{Method \code{new()}}{
This method initializes the "Save" object with optional
 directory argument.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Save$new(wd = getwd())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{wd}}{The directory path for saving files. Default is the
current working directory.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Save-make_filename"></a>}}
\if{latex}{\out{\hypertarget{method-Save-make_filename}{}}}
\subsection{Method \code{make_filename()}}{
This method generates a filename.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Save$make_filename(filename)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{Optional filename string.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The generated filename.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Save-set_dir"></a>}}
\if{latex}{\out{\hypertarget{method-Save-set_dir}{}}}
\subsection{Method \code{set_dir()}}{
This method sets the directory path where files will be saved.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Save$set_dir(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The directory path for saving files.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Save-set_prefix"></a>}}
\if{latex}{\out{\hypertarget{method-Save-set_prefix}{}}}
\subsection{Method \code{set_prefix()}}{
This method sets the prefix used in the file naming.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Save$set_prefix(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The prefix used in the file naming.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Save-set_count"></a>}}
\if{latex}{\out{\hypertarget{method-Save-set_count}{}}}
\subsection{Method \code{set_count()}}{
This method sets the numeric count used in the file naming.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Save$set_count(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The numeric count used in the file naming.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Save-csv"></a>}}
\if{latex}{\out{\hypertarget{method-Save-csv}{}}}
\subsection{Method \code{csv()}}{
This method saves the provided data as a CSV file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Save$csv(data, file, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The data to be saved.}

\item{\code{file}}{The filename for the CSV file.}

\item{\code{...}}{Passed to \code{write.csv}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns the data provided (pipe compatible).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Save-text"></a>}}
\if{latex}{\out{\hypertarget{method-Save-text}{}}}
\subsection{Method \code{text()}}{
This method saves the provided data as a text file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Save$text(data, file, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The data to be saved.}

\item{\code{file}}{The filename for the text file.}

\item{\code{...}}{Passed to \code{sink}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns the data provided (pipe compatible).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Save-plot"></a>}}
\if{latex}{\out{\hypertarget{method-Save-plot}{}}}
\subsection{Method \code{plot()}}{
This method saves the provided plot as an image file in the
 specified device format.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Save$plot(
  plot,
  file,
  device = "png",
  width = 5,
  height = 5,
  units = "in",
  res = 600,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plot}}{The plot to be saved.}

\item{\code{file}}{The filename for the image file.}

\item{\code{device}}{The file type. PDF not currently supported.}

\item{\code{width, height, units, res, ...}}{Sent to device.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns the data provided (pipe compatible).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Save-gg"></a>}}
\if{latex}{\out{\hypertarget{method-Save-gg}{}}}
\subsection{Method \code{gg()}}{
This method saves the provided plot as an image file in the
 specified device format.
 \strong{For gg method you MUST pipe with + instead of \%>\%!}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Save$gg(
  plot,
  file,
  device = "png",
  width = 5,
  height = 5,
  units = "in",
  res = 600,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plot}}{The plot to be saved.}

\item{\code{file}}{The filename for the image file.}

\item{\code{device}}{The file type. PDF not currently supported.}

\item{\code{width, height, units, res, ...}}{Sent to \code{ggsave}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns the data provided (pipe compatible).
 \strong{For gg method you MUST pipe with + instead of \%>\%!}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Save-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Save-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Save$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
