% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.objects.R
\name{savewidget}
\alias{save.objects}
\alias{savewidget}
\title{RStudio Addin to Save Object from Environment to File}
\usage{
save.objects()
}
\description{
RStudio Addin to Save Object from Environment to File
}
\examples{
\dontrun{saveobjects()}
}

